requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
# livekit-server-sdk-python==1.0.2
# livekit-server-sdk==1.0.2
# git+https://github.com/livekit/server-sdk-python.git

python-dotenv==1.0.0
requests==2.31.0